+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ### Overskride Version 0.4.6
 - renaming to duplicate names now adds an index to the dupe (eg: "headphones (1)")
+- auto accept from trusted devices now works
+- added warning for auto accept
+- identified issue where obex agent gets removed after 1 complete transfer
+- fixed issue where a non-identifiable download dir causes a crash